fix: #935 ファイルアップロード時のマジックバイト厳格検証の導入
 ライブラリを使用して、アップロードされたファイルの実際のフォーマット（マジックバイト）をバックエンド側で厳格に検証するようにしました。許可される画像形式を JPEG, PNG, GIF, WEBP に制限し、セキュリティを向上させました。

Closes #935
